verbs v0.5.2 Irregulars

Functions

verb_forms()

Irregular verbs supported by this module:

awake, bear, beat, become, begin, bend, beset, bet, bid, bind, bite, bleed, blow, break, breed, bring, broadcast, build, burn, burst, buy, cast, catch, choose, cling, come, cost, creep, cut, deal, dig, dive, do, draw, dream, drink, drive, eat, fall, feed, feel, fight, find, fit, flee, fling, fly, forbid, forego, forget, forgive, forgo, forsake, freeze, get, give, go, grind, grow, hang, hear, hide, hit, hold, hurt, keep, kneel, knit, know, lay, lead, leap, learn, leave, lend, let, lie, light, lose, make, mean, meet, misspell, mistake, mow, open, overcome, overdo, overtake, overthrow, pay, plead, prove, put, quit, read, reset, rid, ride, ring, rise, run, saw, say, see, seek, sell, send, set, sew, shake, shave, shear, shed, shine, shoe, shoot, show, shrink, shut, sing, sink, sit, slay, sleep, slide, sling, slit, smite, sow, speak, speed, spend, spill, spin, spit, split, spread, spring, stand, steal, stick, sting, stink, stride, strike, string, strive, swear, sweep, swell, swim, swing, take, teach, tear, tell, think, thrive, throw, thrust, tread, understand, uphold, upset, wake, wear, weave, wed, weep, win, wind, withhold, withstand, wring, write

This list may be viewed easily with:

Irregulars.verb_forms |> Map.keys |> Enum.sort |> Enum.join(",") |> IO.puts